test(flex): move the flex fixtures into a conftest so pytest can resolve them
The three fixtures (`capture_backend`, `capture_varlen_backend`, and the autouse
`_reset_backend_overrides`) were defined in `flex_test_utils.py`. That module is a
plain helper -- not a conftest, not a registered plugin -- and no test module
imported the fixture names, so pytest never saw them. On CI this produced 50
`fixture 'capture_backend' not found` errors, and because the autouse reset never
ran either, the module-global override registry leaked across tests and failed two
more in test_flex_routing.py.

Importing the names into each test module would make them visible but every test
taking one as a parameter would then shadow the import (ruff F811), so they move to
`tests/pytorch/ops/conftest.py` instead. The autouse reset is guarded on the module
name: `tests/pytorch/ops` also holds the heavily parametrised kernel suites and
there is no reason to run flex bookkeeping around those.

Also drops a `~` on a bool in test_flex_mask_classify.py (DeprecationWarning today,
an error from Python 3.14).

Verified locally on a CPU-only torch build: 225 passed, both serially and under
`-n 8`; pre-commit --all-files clean.
